+++ src/test/kotlin/net/pwall/json/schema/codegen/CodeGeneratorValidationTest.kt
@@ -1,0 +1,61 @@
+/*
+ * @(#) CodeGeneratorValidationTest.kt
+ *
+ * json-kotlin-schema-codegen  JSON Schema Code Generation
+ * Copyright (c) 2024 Peter Wall
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+package net.pwall.json.schema.codegen
+
+import kotlin.test.Test
+import kotlin.test.assertFailsWith
+import kotlin.test.expect
+
+import java.io.File
+
+import net.pwall.json.schema.JSONSchemaException
+import net.pwall.json.schema.codegen.CodeGeneratorTestUtil.OutputDetails
+import net.pwall.json.schema.codegen.CodeGeneratorTestUtil.outputCapture
+import net.pwall.json.schema.codegen.CodeGeneratorTestUtil.packageDirs
+import net.pwall.json.schema.codegen.CodeGeneratorTestUtil.packageName
+import net.pwall.log.LogList
+import net.pwall.log.assertHasWarning
+
+class CodeGeneratorValidationTest {
+
+    @Test fun `should warn on validation errors`() {
+        LogList().use { logList ->
+            val input = File("src/test/resources/test-validation-errors.schema.json")
+            val outputDetails = OutputDetails(TargetFileName("TestValidationErrors", "kt", packageDirs))
+            CodeGenerator().apply {
+                examplesValidationOption = CodeGenerator.ValidationOption.BLOCK
+                basePackageName = packageName
+                outputResolver = outputCapture(outputDetails)
+                assertFailsWith<JSONSchemaException> { generate(input) }.let {
+                    expect("Validation errors encountered") { it.message }
+                }
+            }
+            logList.assertHasWarning("http://pwall.net/test-validation-errors#/minimum: " +
+                    "Number fails check: minimum 10, was 5, at #/examples/1")
+        }
+    }
+
+}